feat(runloop) add db.connect/close/keepalive to rebuilds
This commit attempts to solve several problems which happen only under heavy load. They are detailed in #4194. This is extracted from #4488

The idea is wrapping every call to the "callback" (update_router or update_plugins_iterator) with calls to kong.db:connect and kong.db:setkeepalive while doing a rebuild.

This seems to have two benefits:
* It solves some sporadic NPEs which seem to occur on timers during heavy load
* It speeds up rebuilds in Cassandra